v8: fix build errors with g++ 6.1.1
Shifting a negative constant value is no longer allowed unless
the -fpermissive flag is in effect.  Fixes the following build
errors:

		../deps/v8/src/objects.h:5188:47: warning: left shift of negative value [-Wshift-negative-value]
			 static const int kElementsKindMask = (-1 << kElementsKindShift) &
		../deps/v8/src/objects.h:5188:44: error: left operand of shift expression '(-1 << 3)' is negative [-fpermissive]
			 static const int kElementsKindMask = (-1 << kElementsKindShift) &
		../deps/v8/src/objects.h:7376:39: warning: left shift of negative value [-Wshift-negative-value]
					 (~kMaxCachedArrayIndexLength << kArrayIndexHashLengthShift) |
		../deps/v8/src/objects.h:7376:36: error: left operand of shift expression '(-8 << 26)' is negative [-fpermissive]
					 (~kMaxCachedArrayIndexLength << kArrayIndexHashLengthShift) |

And:

		../deps/v8/src/liveedit.cc:205:44: warning: left shift of negative value [-Wshift-negative-value]
			 static const int kEmptyCellValue = -1 << kDirectionSizeBits;
		../deps/v8/src/liveedit.cc:205:41: error: left operand of shift expression '(-1 << 2)' is negative [-fpermissive]
			 static const int kEmptyCellValue = -1 << kDirectionSizeBits;

PR-URL: https://github.com/nodejs/node-private/pull/62

+++ deps/v8/src/liveedit.cc
@@ -202,7 +202,8 @@ class Differencer {
 
   static const int kDirectionSizeBits = 2;
   static const int kDirectionMask = (1 << kDirectionSizeBits) - 1;
-  static const int kEmptyCellValue = -1 << kDirectionSizeBits;
+  static const int kEmptyCellValue =
+      static_cast<unsigned int>(-1) << kDirectionSizeBits;
 
   // This method only holds static assert statement (unfortunately you cannot
   // place one in class scope).
