+++ deps/v8/src/objects.h
@@ -5185,7 +5185,8 @@ class Map: public HeapObject {
   static const int kElementsKindBitCount = 5;
 
   // Derived values from bit field 2
-  static const int kElementsKindMask = (-1 << kElementsKindShift) &
+  static const int kElementsKindMask =
+      (static_cast<unsigned int>(-1) << kElementsKindShift) &
       ((1 << (kElementsKindShift + kElementsKindBitCount)) - 1);
   static const int8_t kMaximumBitField2FastElementValue = static_cast<int8_t>(
       (FAST_ELEMENTS + 1) << Map::kElementsKindShift) - 1;
@@ -7373,8 +7374,8 @@ class String: public HeapObject {
   STATIC_CHECK(IS_POWER_OF_TWO(kMaxCachedArrayIndexLength + 1));
 
   static const int kContainsCachedArrayIndexMask =
-      (~kMaxCachedArrayIndexLength << kArrayIndexHashLengthShift) |
-      kIsNotArrayIndexMask;
+      (static_cast<unsigned int>(~kMaxCachedArrayIndexLength)
+          << kArrayIndexHashLengthShift) | kIsNotArrayIndexMask;
 
   // Value of empty hash field indicating that the hash is not computed.
   static const int kEmptyHashField =
